Create new class for title selection code that originally was in SburbHandler
