+++ src/main/java/com/mraof/minestuck/event/ServerEventHandler.java
@@ -13,6 +13,7 @@ import com.mraof.minestuck.player.Echeladder;
 import com.mraof.minestuck.player.IdentifierHandler;
 import com.mraof.minestuck.skaianet.SburbHandler;
 import com.mraof.minestuck.skaianet.SkaianetHandler;
+import com.mraof.minestuck.skaianet.TitleSelectionHook;
 import com.mraof.minestuck.world.MSDimensions;
 import com.mraof.minestuck.world.gen.feature.MSFeatures;
 import com.mraof.minestuck.world.storage.MSExtraData;
@@ -111,7 +112,7 @@ public class ServerEventHandler
 				Echeladder.increaseProgress(player, exp);
 		}
 		if(event.getEntity() instanceof ServerPlayerEntity)
-			SburbHandler.cancelSelection((ServerPlayerEntity) event.getEntity());
+			TitleSelectionHook.cancelSelection((ServerPlayerEntity) event.getEntity());
 	}
 
 	//Gets reset after AttackEntityEvent but before LivingHurtEvent, but is used in determining if it's a critical hit
@@ -168,7 +169,7 @@ public class ServerEventHandler
 	@SubscribeEvent
 	public static void playerChangedDimension(PlayerEvent.PlayerChangedDimensionEvent event)
 	{
-		SburbHandler.cancelSelection((ServerPlayerEntity) event.getPlayer());
+		TitleSelectionHook.cancelSelection((ServerPlayerEntity) event.getPlayer());
 		
 		PlayerSavedData.getData((ServerPlayerEntity) event.getPlayer()).getEcheladder().resendAttributes(event.getPlayer());
 	}
