+++ src/main/java/com/mraof/minestuck/network/TitleSelectPacket.java
@@ -2,7 +2,7 @@ package com.mraof.minestuck.network;
 
 import com.mraof.minestuck.client.gui.MSScreenFactories;
 import com.mraof.minestuck.player.Title;
-import com.mraof.minestuck.skaianet.SburbHandler;
+import com.mraof.minestuck.skaianet.TitleSelectionHook;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.network.PacketBuffer;
 
@@ -47,6 +47,6 @@ public class TitleSelectPacket implements PlayToBothPacket
 	@Override
 	public void execute(ServerPlayerEntity player)
 	{
-		SburbHandler.handleTitleSelection(player, title);
+		TitleSelectionHook.handleTitleSelection(player, title);
 	}
 }
